chore: also add constSysfsExpr to webkit package as it already worked in the transpiler
