+++ typescript-packages/client/src/index.ts
@@ -1,13 +1,14 @@
+export * from './class-mapper';
+export * from './components';
 export * from './custom-components';
 export * from './custom-hooks';
-export * from './components';
 export * from './deck-hooks';
-export * from './modules';
 export * from './globals';
-export * from './webpack';
-export * from './utils';
-export * from './class-mapper';
 export * from './millennium-api';
+export * from './modules';
+export * from './shared/constSysfsExpr';
+export * from './utils';
+export * from './webpack';
 
 import ErrorBoundaryHook from './hooks/error-boundary-hook';
 import RouterHook from './hooks/router/router-hook';
